requirements.txt
streamlit==1.28.1
mistralai==0.0.11
python-dotenv==1.0.0
